New attempt at cross-platform threaded texture loading. Also work on OOMesh implementation based on old Entity drawing code (fancy optimized version having stalled) and fixed problem that was causing synchronization of logcontrol.plist.xml to not happen.
git-svn-id: http://svn.berlios.de/svnroot/repos/oolite-linux/trunk@946 127b21dd-08f5-0310-b4b7-95ae10353056

+++ src/Core/Entities/Entity.h
@@ -223,8 +223,8 @@ extern int debug;
 - (GLfloat *) drawRotationMatrix;
 
 - (BOOL) canCollide;
-- (double) collisionRadius;
-- (void) setCollisionRadius:(double)amount;
+- (GLfloat) collisionRadius;
+- (void) setCollisionRadius:(GLfloat)amount;
 - (NSMutableArray *)collisionArray;
 
 - (void) update:(double) delta_t;
